+++ kits/collectionskit/tests/unit/src/Dictionaries/DictOfNumbersTest.php
@@ -49,6 +49,7 @@ use ReflectionClass;
 use ReflectionMethod;
 use RuntimeException;
 use StusDevKit\CollectionsKit\Dictionaries\CollectionAsDict;
+use StusDevKit\CollectionsKit\Exceptions\NoValueForKeyInCollectionException;
 use StusDevKit\CollectionsKit\Dictionaries\DictOfNumbers;
 
 #[TestDox(DictOfNumbers::class)]
@@ -1082,7 +1083,7 @@ class DictOfNumbersTest extends TestCase
         $this->assertSame(1.99, $actualResult);
     }
 
-    #[TestDox('->get() throws RuntimeException for missing key')]
+    #[TestDox('->get() throws NoValueForKeyInCollectionException for missing key')]
     public function test_get_throws_for_missing_key(): void
     {
         // ----------------------------------------------------------------
@@ -1099,7 +1100,7 @@ class DictOfNumbersTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfNumbers does not contain missing',
         );
@@ -1107,7 +1108,7 @@ class DictOfNumbersTest extends TestCase
         $unit->get('missing');
     }
 
-    #[TestDox('->get() throws RuntimeException for empty dict')]
+    #[TestDox('->get() throws NoValueForKeyInCollectionException for empty dict')]
     public function test_get_throws_for_empty_dict(): void
     {
         // ----------------------------------------------------------------
@@ -1124,7 +1125,7 @@ class DictOfNumbersTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfNumbers does not contain anything',
         );
@@ -1201,7 +1202,7 @@ class DictOfNumbersTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfNumbers does not contain my-special-key',
         );