+++ kits/collectionskit/tests/unit/src/Dictionaries/DictOfStringsTest.php
@@ -48,6 +48,7 @@ use ReflectionClass;
 use ReflectionMethod;
 use RuntimeException;
 use StusDevKit\CollectionsKit\Dictionaries\CollectionAsDict;
+use StusDevKit\CollectionsKit\Exceptions\NoValueForKeyInCollectionException;
 use StusDevKit\CollectionsKit\Dictionaries\DictOfStrings;
 use StusDevKit\CollectionsKit\Traits\StringTransformations;
 
@@ -858,7 +859,7 @@ class DictOfStringsTest extends TestCase
         $this->assertSame('3306', $actualResult);
     }
 
-    #[TestDox('->get() throws RuntimeException for missing key')]
+    #[TestDox('->get() throws NoValueForKeyInCollectionException for missing key')]
     public function test_get_throws_for_missing_key(): void
     {
         // ----------------------------------------------------------------
@@ -875,7 +876,7 @@ class DictOfStringsTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfStrings does not contain missing',
         );
@@ -883,7 +884,7 @@ class DictOfStringsTest extends TestCase
         $unit->get('missing');
     }
 
-    #[TestDox('->get() throws RuntimeException for empty dict')]
+    #[TestDox('->get() throws NoValueForKeyInCollectionException for empty dict')]
     public function test_get_throws_for_empty_dict(): void
     {
         // ----------------------------------------------------------------
@@ -900,7 +901,7 @@ class DictOfStringsTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfStrings does not contain anything',
         );
@@ -951,7 +952,7 @@ class DictOfStringsTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfStrings does not contain my-special-key',
         );