+++ kits/collectionskit/tests/unit/src/Dictionaries/DictOfUuidsTest.php
@@ -52,6 +52,7 @@ use ReflectionNamedType;
 use ReflectionParameter;
 use RuntimeException;
 use StusDevKit\CollectionsKit\Dictionaries\DictOfObjects;
+use StusDevKit\CollectionsKit\Exceptions\NoValueForKeyInCollectionException;
 use StusDevKit\CollectionsKit\Dictionaries\DictOfUuids;
 use StusDevKit\CollectionsKit\Traits\UuidConversions;
 
@@ -858,7 +859,7 @@ class DictOfUuidsTest extends TestCase
         $this->assertSame($uuid2, $actualResult);
     }
 
-    #[TestDox('->get() throws RuntimeException for missing key')]
+    #[TestDox('->get() throws NoValueForKeyInCollectionException for missing key')]
     public function test_get_throws_for_missing_key(): void
     {
         // ----------------------------------------------------------------
@@ -875,7 +876,7 @@ class DictOfUuidsTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfUuids does not contain missing',
         );
@@ -883,7 +884,7 @@ class DictOfUuidsTest extends TestCase
         $unit->get('missing');
     }
 
-    #[TestDox('->get() throws RuntimeException for empty dict')]
+    #[TestDox('->get() throws NoValueForKeyInCollectionException for empty dict')]
     public function test_get_throws_for_empty_dict(): void
     {
         // ----------------------------------------------------------------
@@ -900,7 +901,7 @@ class DictOfUuidsTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfUuids does not contain anything',
         );
@@ -952,7 +953,7 @@ class DictOfUuidsTest extends TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(NoValueForKeyInCollectionException::class);
         $this->expectExceptionMessage(
             'DictOfUuids does not contain my-special-key',
         );