+++ CHANGELOG.md
@@ -22,7 +22,7 @@ This CHANGELOG is automatically compiled.
 
 #### CollectionsKit
 
-- added NoValueForKeyInCollectionException ([2844a44](https://github.com/stuartherbert/stusdevkit/commit/2844a44)) — Stuart Herbert <stuart@stuartherbert.com>
+- added NoValueForKeyInCollectionException ([c1fd4de](https://github.com/stuartherbert/stusdevkit/commit/c1fd4de)) — Stuart Herbert <stuart@stuartherbert.com>
 - added CollectionAsStack and StackOfStrings ([23db293](https://github.com/stuartherbert/stusdevkit/commit/23db293)) — Stuart Herbert <stuart@stuartherbert.com>
 
 #### MissingBitsKit
@@ -173,6 +173,7 @@ This CHANGELOG is automatically compiled.
 
 #### CollectionsKit
 
+- CollectionAsDict::get() now throws NoValueForKeyInCollectionException ([8ec6e76](https://github.com/stuartherbert/stusdevkit/commit/8ec6e76)) — Stuart Herbert <stuart@stuartherbert.com>
 - move AccessibleCollection::copy() to CollectionOfAnything::copy() ([c7ad3c3](https://github.com/stuartherbert/stusdevkit/commit/c7ad3c3)) — Stuart Herbert <stuart@stuartherbert.com>
 - moved Arrayable into MissingBitsKit ([f80dc1c](https://github.com/stuartherbert/stusdevkit/commit/f80dc1c)) — Stuart Herbert <stuart@stuartherbert.com>
 - introduce AccessibleCollection intermediate class ([565e5fe](https://github.com/stuartherbert/stusdevkit/commit/565e5fe)) — Stuart Herbert <stuart@stuartherbert.com>